Meeting Requests for PR
Basically making it more customizable

"death_buff" is a bool that if true makes it so that the creature gets the Ouroboro's trait.

"frozen_card" is a field for a card's name; if no evolution for a creature with is specified then it releases the card specified in "frozen_card". if no card is specified in "frozen card" and evolution specified then it is an Opossum.

The Mental Gemnastics fix actually fixes what it was intended to fix now.