Final? Pass
Misc changes to comments to make them clearer, as well as cleaning up some unnecessary additions that seemed necessary at the time.

Gem Guardian actually adds the shielded sigil to the card. This is neccessary because the sigil does the shield logic now, not the effect.

Gem Animator is no longer hardcoded. I honestly kept ignoring this one throughout the entire time for reasons unknown to even myself.